added CallerMemberName support
When applying an XorProperty or XorReference attribute to a property, it
is no longer required to specify a name for it. The C# compiler will
automatically do this for you, using the property name.
It seems that the feature does not work on fields.

Added InvalidXorAttributeNameException. Added tests for the feature.
Extended the README.md with a short description of the feature. Modified
the example.